perf(desktop): stand the voice mark still, and stop paying for it every frame
Measured in the running app through CDP, not a standalone harness — the
harness had said the canvas was fine while the app stuttered. Voice Mode
cost 19-40% of this 120Hz machine's CPU/GPU against 1.2% for the chat
surface; Focus Mode on top of it passed 50%. Repeated tight A/Bs put ~10
points on the animated hologram, ~11 on four 2px bars in the control dock,
and ~15 on the Focus spectrogram.

So the mark stops turning. It stands where the app's own logo stands and
is painted once — on resize, theme change, phase colour, or a change in
the running operations. What moves with the voice is one CSS custom
property published fifteen times a second, driving a scale and an opacity.
Both are compositor properties: they animate without a repaint, and they
keep animating while the main thread is busy, which is precisely what the
old canvas loop could not do when the renderer was saturated. The loop
parks once the level settles with no analyser feeding it, so idle Voice
Mode costs nothing and a headless renderer cannot spin a frame loop.

The dock bars now declare both keyframe ends. A set with only a `to` leaves
the start value to be inferred from computed style, and an animation with
an implicit keyframe cannot be handed to the compositor — which is why four
2px bars were forcing a style recalculation on every frame of a 120Hz
display. The Focus spectrogram is paced to 30Hz and drops its per-bar
canvas shadow, redundant against the 10px CSS blur the element already has.

Verified in the app afterwards: Voice Mode 3.0% with zero style
recalculations in six seconds, Focus Mode 0.0%, the mark rendering with
both strands and the pulse driving the scale exactly as specified. 642
desktop tests, build, typecheck and lint green.

Not fixed, and now the largest remaining cost: during an active turn the
renderer main thread runs 92% busy with 53% of the window in JavaScript,
while style recalculation stays at the pulse's 15Hz. That is the transcript
and React render path, not the graphics.
