+++ apps/desktop/src/voice-call/VoiceHologram.tsx
@@ -1,5 +1,6 @@
 import type { VoiceCallPhase } from '@moxxy/client-core';
-import { useVoiceHologramAnimation } from './useVoiceHologramAnimation';
+import { useVoiceHologramScene } from './useVoiceHologramScene';
+import { useVoicePulse } from './useVoicePulse';
 
 export function VoiceHologram({
   phase,
@@ -14,18 +15,17 @@ export function VoiceHologram({
   readonly outputAnalyser: unknown | null;
   readonly occupiedSlots: ReadonlyArray<number>;
 }): JSX.Element {
-  const canvasRef = useVoiceHologramAnimation({
-    phase,
-    inputAnalyser,
-    outputAnalyser,
-    occupiedSlots,
-  });
+  const canvasRef = useVoiceHologramScene({ phase, occupiedSlots });
+  const pulseRef = useVoicePulse({ phase, inputAnalyser, outputAnalyser });
+
   return (
     <div
+      ref={pulseRef}
       className={`voice-hologram voice-hologram--${phase}${microphoneMuted ? ' is-muted' : ''}`}
       aria-hidden="true"
     >
       <canvas ref={canvasRef} className="voice-hologram-canvas" data-testid="voice-hologram-canvas" />
+      <span className="voice-hologram-pulse" />
     </div>
   );
 }
